delete swap action, make it mutation
